backend/requirements.txt
fastapi==0.115.8
ollama==0.4.7
pydantic==2.10.6
python-dotenv==1.0.1
uvicorn==0.34.0